+++ android/src/main/kotlin/net/touchcapture/qr/flutterqr/QRView.kt
@@ -5,48 +5,39 @@ import android.app.Activity
 import android.app.Application
 import android.content.Context
 import android.content.pm.PackageManager
-import android.content.pm.PackageManager.PERMISSION_GRANTED
-import android.os.Build
 import android.os.Bundle
 import android.view.View
 import com.google.zxing.ResultPoint
 import android.hardware.Camera.CameraInfo
+import android.os.Build
 import com.journeyapps.barcodescanner.BarcodeCallback
 import com.journeyapps.barcodescanner.BarcodeResult
 import com.journeyapps.barcodescanner.BarcodeView
+import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
-import io.flutter.plugin.common.PluginRegistry
 import io.flutter.plugin.platform.PlatformView
+class QRView(messenger: BinaryMessenger, id: Int, private val context: Context) :
+        PlatformView, MethodChannel.MethodCallHandler {
 
-class QRView(private val registrar: PluginRegistry.Registrar, id: Int) :
-        PlatformView,MethodChannel.MethodCallHandler {
-
-    companion object {
-        const val CAMERA_REQUEST_ID = 513469796
-    }
-
-    var barcodeView: BarcodeView? = null
-    private val activity = registrar.activity()
-    var cameraPermissionContinuation: Runnable? = null
-    var requestingPermission = false
     private var isTorchOn: Boolean = false
-    val channel: MethodChannel
+    private var barcodeView: BarcodeView? = null
+    private var requestingPermission = false
+    private val channel: MethodChannel
 
     init {
-        registrar.addRequestPermissionsResultListener(CameraRequestPermissionsListener())
-        channel = MethodChannel(registrar.messenger(), "net.touchcapture.qr.flutterqr/qrview_$id")
-        channel.setMethodCallHandler(this)
         checkAndRequestPermission(null)
-        registrar.activity().application.registerActivityLifecycleCallbacks(object : Application.ActivityLifecycleCallbacks {
+        channel = MethodChannel(messenger, "net.touchcapture.qr.flutterqr/qrview_$id")
+        channel.setMethodCallHandler(this)
+        Shared.activity?.application?.registerActivityLifecycleCallbacks(object : Application.ActivityLifecycleCallbacks {
             override fun onActivityPaused(p0: Activity) {
-                if (p0 == registrar.activity()) {
+                if (p0 == Shared.activity) {
                     barcodeView?.pause()
                 }
             }
 
             override fun onActivityResumed(p0: Activity) {
-                if (p0 == registrar.activity()) {
+                if (p0 == Shared.activity) {
                     barcodeView?.resume()
                 }
             }
@@ -68,12 +59,34 @@ class QRView(private val registrar: PluginRegistry.Registrar, id: Int) :
         })
     }
 
-    fun flipCamera() {
+    override fun dispose() {
         barcodeView?.pause()
-        var settings = barcodeView?.cameraSettings
+        barcodeView = null
+    }
+
+    override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
+        when(call.method){
+            "flipCamera" -> {
+                flipCamera()
+            }
+            "toggleFlash" -> {
+                toggleFlash()
+            }
+            "pauseCamera" -> {
+                pauseCamera()
+            }
+            "resumeCamera" -> {
+                resumeCamera()
+            }
+        }
+    }
+
+    private fun flipCamera() {
+        barcodeView?.pause()
+        val settings = barcodeView?.cameraSettings
 
         if(settings?.requestedCameraId == CameraInfo.CAMERA_FACING_FRONT)
-            settings?.requestedCameraId = CameraInfo.CAMERA_FACING_BACK
+            settings.requestedCameraId = CameraInfo.CAMERA_FACING_BACK
         else
             settings?.requestedCameraId = CameraInfo.CAMERA_FACING_FRONT
 
@@ -102,7 +115,7 @@ class QRView(private val registrar: PluginRegistry.Registrar, id: Int) :
     }
 
     private fun hasFlash(): Boolean {
-        return registrar.activeContext().packageManager
+        return context.packageManager
                 .hasSystemFeature(PackageManager.FEATURE_CAMERA_FLASH)
     }
 
@@ -119,12 +132,16 @@ class QRView(private val registrar: PluginRegistry.Registrar, id: Int) :
         return barcodeView
     }
 
-    private fun createBarCodeView(): BarcodeView? {
-        val barcode = BarcodeView(registrar.activity())
+    private fun createBarCodeView(): BarcodeView {
+        val barcode = BarcodeView(Shared.activity)
         barcode.decodeContinuous(
                 object : BarcodeCallback {
                     override fun barcodeResult(result: BarcodeResult) {
-                        channel.invokeMethod("onRecognizeQR", result.text)
+                        val code = mapOf(
+                                "code" to result.text,
+                                "type" to result.barcodeFormat.name,
+                                "rawBytes" to result.rawBytes)
+                        channel.invokeMethod("onRecognizeQR", code)
                     }
 
                     override fun possibleResultPoints(resultPoints: List<ResultPoint>) {}
@@ -133,54 +150,18 @@ class QRView(private val registrar: PluginRegistry.Registrar, id: Int) :
         return barcode
     }
 
-    override fun dispose() {
-        barcodeView?.pause()
-        barcodeView = null
-    }
-
-    private inner class CameraRequestPermissionsListener : PluginRegistry.RequestPermissionsResultListener {
-        override fun onRequestPermissionsResult(id: Int, permissions: Array<String>, grantResults: IntArray): Boolean {
-            if (id == CAMERA_REQUEST_ID && grantResults[0] == PERMISSION_GRANTED) {
-                cameraPermissionContinuation?.run()
-                return true
-            }
-            return false
-        }
-    }
-
     private fun hasCameraPermission(): Boolean {
         return Build.VERSION.SDK_INT < Build.VERSION_CODES.M ||
-                activity.checkSelfPermission(Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED
-    }
-
-
-    override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
-        when(call?.method){
-            "checkAndRequestPermission" -> {
-                checkAndRequestPermission(result)
-            }
-            "flipCamera" -> {
-                flipCamera()
-            }
-            "toggleFlash" -> {
-                toggleFlash()
-            }
-            "pauseCamera" -> {
-                pauseCamera()
-            }
-            "resumeCamera" -> {
-                resumeCamera()
-            }
-        }
+                Shared.activity?.checkSelfPermission(Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED
     }
 
     private fun checkAndRequestPermission(result: MethodChannel.Result?) {
-        if (cameraPermissionContinuation != null) {
-            result?.error("cameraPermission", "Camera permission request ongoing", null);
+        if (Shared.cameraPermissionContinuation != null) {
+            result?.error("cameraPermission", "Camera permission request ongoing", null)
         }
 
-        cameraPermissionContinuation = Runnable {
-            cameraPermissionContinuation = null
+        Shared.cameraPermissionContinuation = Runnable {
+            Shared.cameraPermissionContinuation = null
             if (!hasCameraPermission()) {
                 result?.error(
                         "cameraPermission", "MediaRecorderCamera permission not granted", null)
@@ -190,17 +171,27 @@ class QRView(private val registrar: PluginRegistry.Registrar, id: Int) :
 
         requestingPermission = false
         if (hasCameraPermission()) {
-            cameraPermissionContinuation?.run()
+            Shared.cameraPermissionContinuation?.run()
         } else {
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
                 requestingPermission = true
-                registrar
-                        .activity()
-                        .requestPermissions(
-                                arrayOf(Manifest.permission.CAMERA),
-                                CAMERA_REQUEST_ID)
+                Shared.activity?.requestPermissions(
+                        arrayOf(Manifest.permission.CAMERA),
+                        Shared.CAMERA_REQUEST_ID)
             }
         }
     }
 
+    override fun onFlutterViewAttached(flutterView: View) {
+    }
+
+    override fun onFlutterViewDetached() {
+    }
+
+    override fun onInputConnectionLocked() {
+    }
+
+    override fun onInputConnectionUnlocked() {
+    }
 }
+