+++ android/src/main/kotlin/net/touchcapture/qr/flutterqr/Shared.kt
@@ -1,0 +1,9 @@
+package net.touchcapture.qr.flutterqr
+
+import android.app.Activity
+
+object Shared {
+    const val CAMERA_REQUEST_ID = 513469796
+    var activity: Activity? = null
+    var cameraPermissionContinuation: Runnable? = null
+}