+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.2.0
+#### Breaking change
+* The plugin now returns Barcode object instead of QR String. This object includes the type of code, the code itself and on Android devices the raw bytes. (#63)
+#### New Features
+* Added possibility to provide scanArea on iOS. (#165)
+#### Bug fixes
+* Fixed preview going black after hot reload. (#76)
+* Fixed nullpointer when plugin binding order isn't correct. (#181)
+* Fixed permission being asked on startup (#185)
+
+## 0.1.0
+* Changed Android minSDKversion from 24 to 21 (#170)
+* Fix preview size after iPad rotation (#125)
+* Implemented Android Embedding V2 (#132)
+* Added cutout bottom offset (#115)
+* Fix Android ActivityLifecycleCallbacks (#166)
+* Fix some other small bugs
+
+## 0.0.14
+* Fix disposing camera on iOS 14 (#113)
+
 ## 0.0.13
 * Fix misalignment when QRView doesn't start from the top left (#45)
 * Fix crash on iOS when scanning returns nil (#69, #72)